tests: Fix dtype mismatches after DTYPE-DEFAULT-001 migration
Fixed 27+ test failures caused by dtype incompatibility between test
expectations (float64) and implementation defaults (float32).

Changes:
- Add `dtype=torch.float64` to Detector/Crystal/Simulator constructors
  in precision-critical tests (geometry, gradients, basis vectors)
- Updated 11 test files to explicitly request float64 where needed per
  arch.md §313 guidance ("Precision-critical operations override to
  float64 explicitly where required")

Test Results:
- test_at_geo_003.py: 8/8 passed
- test_at_geo_004.py: 6/6 passed
- test_detector_basis_vectors.py: 7/7 passed
- test_at_parallel_017.py: 6/6 passed
- Overall: 27+ tests fixed, reducing total failures from 36 to ~9

Remaining dtype issues in IO module (read_hkl_file, read_sourcefile)
require separate fix to pass dtype parameter to IO functions.

Refs: TEST-DTYPE-001 in docs/fix_plan.md

+++ tests/test_at_geo_004.py
@@ -76,7 +76,7 @@ def test_twotheta_rotation_applied():
         beam_center_f=51.2
     )
 
-    detector = Detector(config)
+    detector = Detector(config, dtype=torch.float64)
 
     # The twotheta rotation should affect the detector basis vectors
     # For XDS with twotheta axis [1,0,0], rotation should be around X axis
@@ -122,7 +122,7 @@ def test_twotheta_value_preserved():
         assert config.detector_twotheta_deg == test_twotheta_deg
 
         # Create detector and verify the value is used
-        detector = Detector(config)
+        detector = Detector(config, dtype=torch.float64)
         # The detector internally converts to radians but the config value should remain in degrees
         assert config.detector_twotheta_deg == test_twotheta_deg
 
@@ -137,7 +137,7 @@ def test_mosflm_twotheta_rotation():
         pixel_size_mm=0.1
     )
 
-    detector = Detector(config)
+    detector = Detector(config, dtype=torch.float64)
 
     # For MOSFLM with twotheta axis [0,0,-1], rotation should be around negative Z axis
     # Original MOSFLM basis: fdet=[0,0,1], sdet=[0,-1,0], odet=[1,0,0]
